refactor(client): extract chat-header widgets to dedicated file
The header file had 4 private widget classes living inside it
(IdentityChangedBadge, VerifiedBadge, TimerChip, PinnedMessagesDialog),
which inflated chat_header_bar.dart to 1254 LOC. Move all four to a
new chat_header_widgets.dart and drop the leading underscore so they
can be imported.

Pure mechanical refactor — no logic changes, no API changes. Existing
tests pass unchanged. chat_header_bar.dart drops to 805 LOC (-36%);
the move also lets us remove three now-unused imports
(dart:convert is still used for the safety-number flow, so it stays).
